Setup basic cpuid

+++ src/utils/bitutils.rs
@@ -1,3 +1,3 @@
-pub fn is_nth_bit_set(number: u8, bit_index: usize) -> bool {
+pub fn is_nth_bit_set(number: usize, bit_index: usize) -> bool {
     number & (1 << bit_index) != 0
 }
